Add JsonPath, TypeParser, and resource files for more Pokémon.
JsonPath was added to build.gradle, along with code which addresses an
error (Dr. G provided this code during the Two-Week Project).

TypeParser has been tested and is able to parse through a string of data in
JSON format taken from the PokeAPI to find the type(s) of a Pokémon or the
Pokémon's original type(s) if it has changed.