Restructure TypeParser to adhere to SRP, add PokemonParser and data classes.
TypeParser was broken up into several more specific classes. The Pokemon
and Type data structures were created, along with builders for each.

A few more redundant tests were removed, and a few more json resources were
added for future testing.

There is unfinished code in PokemonBuilder which, when completed, will be
able to accurately return the types which a Pokemon with multiple types is
weak, resistant, and immune to.