Refine how TypeBuilder and Parser get damage relations of a type.
Previously, a string called jsonPathRoot was created and passed into
several methods to assist with parsing the entire type json object.

I realized that this could be replaced by instead having TypeBuilder pass
the LinkedHashMap of the damage relations into TypeParser, so that
TypeParser does not have to parse the entire type json object.

This also means that the string jsonPathRoot is no longer needed, meaning
less parameters are being passed between classes.