Refine how stats are parsed and added to the Pokemon.
The parsing for all stats is now handled by a single, more generic,
function in PokemonParser.

The new method also means that PokemonBuilder receives a HashMap from
parser, so it no longer needs to build one.

JsonParser gained a new method so that it can convert a JSONArray into a
List of Integers.