Create MoveBuilder and add parsers to PokemonParser for creating moves.
Had to address a few unique problems, such as the fact that some moves
can be learned multiple times by a Pokemon (see Leer and Charizard).

Due to this revelation, had to change learnMethod into a list. Also added
the list levelLearned to the Move data structure.

PokemonParser gets all relevant information about a move that it can from
the json data we get for a Pokemon.