Add in commented out code for caching JSON Objects.
The slowest part of our program is, by far, the amount of times we have to
connect to the internet.

I noticed that some of these calls were redundant, such as pokedexes and
moves.

Added code which caches JSON Objects and links them to URLs, so that if
the URL is called again, a connection is not made and instead the program
just returns the cached JSON Object.

The main downside to this method is that it takes up memory. Worth doing
more investigating.