Refactor parsers to make better use of JSONPath and address XD issue.
Parsers used to default to getting JSONArrays for everything, but now
they get more specific types of objects when possible (such as integers and
strings).

Also, changed code in PokedexParser so that if a game does not have a
pokedex, such as XD and Colosseum, an empty string list is returned.
Previously, a null value was returned, which resulted in the wrong
error being displayed to the user.