Restructure how list of all versions is created and retrieved.
Instead of getting a list of generations, and interacting with that, we can
just fetch a list of versions, which is created by first getting a list of
generations and then version groups.

Parsers were restructured so that they no longer contain any URL calls
or connection to the internet. All connection to internet will be handled
in Fetcher or Processor classes.

Slight modifications were made to view, just to account for the change from
a list of generations to a list of versions.

All Parers are now thoroughly tested.