Refactor Pokedex into NationalPokedex.
After talking with Dr. G, I realized that NationalPokedexFetcher was not doing much
and that Pokedex was obsolete. Converted Pokedex into NationalPokedex and gave it a
Factory class that it can create itself when it is called.

Also, moved URL into URLProcessor. I think it would be wise if we keep all of our
URLs in URLProcessor, to make it easier to refactor if PokeAPI changes their URL.