Refactor Version to add GenerationMap attribute.
GenerationMap allows the parsers to, through a Version object, figure out whether a
generation came before or after another generation. Useful when parsing for past
types and stats.

Unfortunately, GenerationMap is tough to create locally, which means that the test
classes for PokemonEngineer and PokemonParser are communicating to the internet when
they run, making the tests quite a bit slower.

This needs to be refactored further to get the tests to no longer be dependent on the
internet, but for now everything passes, just slowly.