Refactor Pokedex into NationalPokedex and Refactor Version to add GenerationMap attribute
* Refactor Pokedex into NationalPokedex.

After talking with Dr. G, I realized that NationalPokedexFetcher was not doing much
and that Pokedex was obsolete. Converted Pokedex into NationalPokedex and gave it a
Factory class that it can create itself when it is called.

Also, moved URL into URLProcessor. I think it would be wise if we keep all of our
URLs in URLProcessor, to make it easier to refactor if PokeAPI changes their URL.

* Refactor Version to add GenerationMap attribute.

GenerationMap allows the parsers to, through a Version object, figure out whether a
generation came before or after another generation. Useful when parsing for past
types and stats.

Unfortunately, GenerationMap is tough to create locally, which means that the test
classes for PokemonEngineer and PokemonParser are communicating to the internet when
they run, making the tests quite a bit slower.

This needs to be refactored further to get the tests to no longer be dependent on the
internet, but for now everything passes, just slowly.