Implement URLProcessor so that there is a test and production URLProcessor
TestURLProcessor does not actually connect to the internet, and can be used to replace
ProductionURLProcessor when running tests so that we are not connecting to the internet.

Dr. G recommended this approach so that we can test more of our code. He also said that
as long as the Test versions of URL Processor is in the Test directory, then it is
good practice.