Refactor classes and tests so that no tests have to connect to the API.
Had to rename some resource files unfortunately in order to allow TestURLProcessor to
convert the URL it gets in a document into a json document. We can refactor this later
if we want.

Deleted TestsWithVersions as it is no longer needed, and all it did was connect to the
API.

Some classes had extra constructors created so that dependency injection could occur.

Added TODO to PokemonEngineer as I do not believe it is needed, and can be replaced with
Pokemon.Builder