Refactor PokemonParser and MoveDisplay to retrieve and display correct moves.
Game indices stopped being kept track of by PokeAPI after Black and White 2, which
means a new method had to be used to determine if a Pokémon is in a game.

This new method actually just needs the Version Group, meaning we might be able to
delete the Version class and just replace it with Version Group.

Added TODO to MoveDisplay because it does not account for moves with multiple learn
methods.