Add PokemonSpeciesParser to parse for the url of the default version of a Pokemon
Some Pokemon, such as Giratina, are in the National Pokedex, but they have different
forms, so their url is not "/pokemon/giratina" but instead something else.

In order to determine what this url is, it is necessary to go to the pokemon-species
page first and find the default form's url.