Refactor test in EvolutionChainParserTest to be more thorough and accurately named
As we learned in class, it is okay for a test to have multiple assertions if those
assertions are not testing different things. In this scenario, since we are testing
the hashmap of evolution details, I think that it makes sense we assert that each value
in the hashmap is what we expect for the charmeleon evolution.