Add move cache to PokemonParser to speed up time it takes to get Pokémon forms
A lot of Pokémon forms share similar, if not identical, moves to the default form. Added
a HashMap which contains the url of a move as its key, and the Move object as its value.

This solution seems to be light on memory usage, since we are just storing the Move
object instead of the entire Move json document. Additionally, the cache is wiped every
time a new Pokémon is searched.

Unfortunately, this does mean that parseForMoves now has three parameters, but I believe
this is justifiable given how slow the program is without a cache.