Refactor to address bug in MoveDisplay and optimize moveCache in PokemonParser
MoveDisplay was not displaying every move a Pokémon had, skipping its first move.
Noticed this when looking up Kakuna, which was displaying no moves.

Instead of storing the entire URL in the hashmap for moveCache, now we just store
the last part of the URL which includes the move's unique id. This should make it a bit
faster when dealing with Pokémon which have a lot of moves.