Refactor LinkedHashMap of Evolution Details to make it more specific and address error
Previously, we implemented the hashmap of evolution details in the jsonDocument as a
LinkedHashMap<String, Object> but Object is too broad, especially since we know that the
values can only be boolean, integer, string, or LinkedHashMap<String, String>.

Created the EvolutionDetailsValues interface and classes which implement it so that
we can have a more specific LinkedHashMap<String, EvolutionDetailsValues>. This
also means that our view layer no longer has to do any manipulations to get the
name from a LinkedHashMap value, keeping the view humble.

Also made sure that trigger is always the first item in the LinkedHashMap
EvolutionDetails since that is what the view layer expects.

Unfortunately, it appears that we cannot safely cast a value to an instance of
LinkedHashMap<String, String> so we just cast it to a LinkedHashMap<?,?> in
EvolutionDetailLinkedHashMap.