Add in multiple new data structures to more cleanly deal with Evolution Details
After talking to Dr. G, he recommended we convert the data we receive from JsonPath into
a domain model, so that our domain model is not dependent on the way the data is
stored in PokeAPI.

This required a lot of "boilerplate" code unfortunately, and this may end up not
being doable in the time we have left.

Still though, it does not hurt the program to include these classes for now. They are
not called when the application is run, but if we find use for them, they should be
easy to incorporate.

There are some warnings listed because not all the methods are used in the data
structures, however all of these methods will be necessary if we do fully incorporate
these classes into the application. Figured it is better to have those methods ready.